NXP-23125: move tests related to queries with tags to tag module
